requirements.txt
gql==0.1.0
graphql-core==2.1
promise==2.1
Rx==1.6.1
six==1.11.0
typing==3.6.4
click
mutagen
boto3
PyYAML
graphqlclient